When `State` is implemented with `sealed classes`, `StateCompletable` was forcing the developer to create an "end state" case in the `sealed class` to terminate the flow. In some cases, e.g. the `render` method, you will be forced to take care of this unreachable case. In fact, the `render` will never be called when the flow ends, aka when the `State` becomes that "end case".
`StateCompletabe` was removed.

`Stepper` instead of `Reducer`:
`(State, Event) -> State` forces us to produce new `State` but doesn't give us a way to end the flow.
The concept was modified to return `Step` instead of a new `State`.
A `Step` be either a new `State` or an `Output` of the flow.
When a new state is returned the flow will be advancing but when an `Output` is returned the flow will immediately complete.