Added the concept of flow id that would allow to identify a flow per input.
When context.renderChild is called, the `childFlow` is started with provided `input` if the `childFlow.id(input)` is not already started. Otherwise, the flow is resumed.

Moved the ViewModal concept outside the feedbacktree core library. It is now part of the sample app. Other developers can be inspired by this code and implement their own solution.